app/requirements.txt
streamlit==1.17.0
honeybee-display==0.2.17
ladybug-charts==1.18.8
ladybug-vtk==0.13.9
pollination-streamlit==0.5.0
pollination_streamlit_io>=0.49.0
pollination_streamlit_viewer==0.4.7
honeybee-radiance==1.64.150
